// packages/signal-ts/src/client/protos/compiled.d.ts
export namespace signalservice {
// ...
    namespace Envelope {

        /** Type enum. */
        enum Type {
            UNKNOWN = 0,
            CIPHERTEXT = 1,
            KEY_EXCHANGE = 2,
            PREKEY_BUNDLE = 3,
            RECEIPT = 5,
            UNIDENTIFIED_SENDER = 6,
            PLAINTEXT_CONTENT = 8
        }
    }
// ...
    namespace DataMessage {
// ...
        namespace Contact {

            /** Properties of a Name. */
            interface IName {

                /** Name givenName */
                givenName?: (string|null);

                /** Name familyName */
                familyName?: (string|null);

                /** Name prefix */
                prefix?: (string|null);

                /** Name suffix */
                suffix?: (string|null);

                /** Name middleName */
                middleName?: (string|null);

                /** Name displayName */
                displayName?: (string|null);
            }
// ...
            interface IPhone {

                /** Phone value */
                value?: (string|null);

                /** Phone type */
                type?: (signalservice.DataMessage.Contact.Phone.Type|null);

                /** Phone label */
                label?: (string|null);
            }
// ...
            interface IEmail {

                /** Email value */
                value?: (string|null);

                /** Email type */
                type?: (signalservice.DataMessage.Contact.Email.Type|null);

                /** Email label */
                label?: (string|null);
            }
// ...
            interface IPostalAddress {

                /** PostalAddress type */
                type?: (signalservice.DataMessage.Contact.PostalAddress.Type|null);

                /** PostalAddress label */
                label?: (string|null);

                /** PostalAddress street */
                street?: (string|null);

                /** PostalAddress pobox */
                pobox?: (string|null);

                /** PostalAddress neighborhood */
                neighborhood?: (string|null);

                /** PostalAddress city */
                city?: (string|null);

                /** PostalAddress region */
                region?: (string|null);

                /** PostalAddress postcode */
                postcode?: (string|null);

                /** PostalAddress country */
                country?: (string|null);
            }
// ...
        }
// ...
    }
// ...
}